# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "API endpoints for managing hotel bookings",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "operationId": "30ce4124bc07a0d013b69974ff225a5c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerWithId"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "8350a1a36d2b5522ea4158253eff28a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerPOST"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer by ID",
                "operationId": "bb85afd644c1b965544c2f0e51f408b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerWithId"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer by ID",
                "operationId": "7f856cee51366cc0a4c5e1aa6a02547a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerPUT"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer by ID",
                "operationId": "4cee47d6b826da6ed7d852b5aff46d19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Customer deleted successfully"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerPOST": {
                "title": "Customer for POST",
                "description": "Customer object",
                "required": [
                    "name",
                    "email",
                    "phone_number"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the customer",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the customer",
                        "type": "string",
                        "example": "johnUpdated@example.com"
                    },
                    "phone_number": {
                        "description": "Phone number of the customer",
                        "type": "string",
                        "example": 1234567890
                    }
                },
                "type": "object"
            },
            "CustomerPUT": {
                "title": "Customer for PUT",
                "description": "Customer object",
                "properties": {
                    "name": {
                        "description": "Name of the customer",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the customer",
                        "type": "string",
                        "example": "johnUpdated@example.com"
                    },
                    "phone_number": {
                        "description": "Phone number of the customer",
                        "type": "string",
                        "example": 1234567890
                    }
                },
                "type": "object"
            },
            "CustomerWithId": {
                "title": "Customer with ID",
                "description": "Customer object",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the customer",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the customer",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of the customer",
                        "type": "string",
                        "example": "johnUpdated@example.com"
                    },
                    "phone_number": {
                        "description": "Phone number of the customer",
                        "type": "string",
                        "example": 1234567890
                    },
                    "created_at": {
                        "description": "Timestamp when the booking was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-13 12:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp when the booking was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-13 12:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    }
}